components: Make KeyboardHandler importable
